create model inference smp in my project

+++ segmentation/scr/create_tilling_data.py
@@ -14,7 +14,8 @@ from segmentation.config import CFG
 from collections import defaultdict
 from segmentation.scr.utils.rle_coding import rle_decode, rle_encode
 
-warnings.filterwarnings("ignore", category=rasterio.errors.NotGeoreferencedWarning)
+warnings.filterwarnings(
+    "ignore", category=rasterio.errors.NotGeoreferencedWarning)
 
 
 @lru_cache(maxsize=128)
@@ -79,8 +80,10 @@ class Generate_Tiled_Dataset(Dataset):
             images_labels[img.name].append(img)
         for lb in self.path_lb_dir:
             images_labels[lb.name].append(lb)
-        new_dict = dict(filter(lambda item: len(item[1]) > 1, images_labels.items()))
-        print(f"Общее число изображений с масками сегментации : {len(new_dict)}")
+        new_dict = dict(filter(lambda item: len(
+            item[1]) > 1, images_labels.items()))
+        print(
+            f"Общее число изображений с масками сегментации : {len(new_dict)}")
 
         self.samples = []
         for key in tqdm(new_dict, total=len(new_dict), desc="Reading images"):
@@ -91,7 +94,8 @@ class Generate_Tiled_Dataset(Dataset):
                 img_r = reader.read()
                 px_max, px_min = img_r.max(), img_r.min()
             # print(p_lb)
-            self.samples.append((p_img, p_lb, [px_min, px_max], (height, width)))
+            self.samples.append(
+                (p_img, p_lb, [px_min, px_max], (height, width)))
 
         min_overlap = float(overlap_pct) * 0.01
         max_stride = self.tile_size * (1.0 - min_overlap)
@@ -105,15 +109,17 @@ class Generate_Tiled_Dataset(Dataset):
         ):
             # [(C,H,W),...]
             height, width = img_dims
-            num_patches = np.ceil(np.array([height, width]) / max_stride).astype(
-                np.int64
-            )
+            num_patches = np.ceil(np.array([height, width]) / max_stride)
             starts = [
-                np.int32(np.linspace(0, height - self.tile_size[0], num_patches[0])),
-                np.int32(np.linspace(0, width - self.tile_size[1], num_patches[1])),
+                np.int32(np.linspace(
+                    0, height - self.tile_size[0], num_patches[0])),
+                np.int32(np.linspace(
+                    0, width - self.tile_size[1], num_patches[1])),
             ]
-            stops = [starts[0] + self.tile_size[0], starts[1] + self.tile_size[1]]
-            mask = cv2.imread(str(label_path), cv2.IMREAD_GRAYSCALE).astype(np.uint8)
+            stops = [starts[0] + self.tile_size[0],
+                     starts[1] + self.tile_size[1]]
+            mask = cv2.imread(
+                str(label_path), cv2.IMREAD_GRAYSCALE).astype(np.uint8)
             rle = rle_encode(mask)
 
             for y1, y2 in zip(starts[0], stops[0]):
@@ -140,7 +146,8 @@ class Generate_Tiled_Dataset(Dataset):
                             (height, width),
                         )
                     )
-        print(f"Dataset contains {empty} empty and {nonempty} non-empty tiles.")
+        print(
+            f"Dataset contains {empty} empty and {nonempty} non-empty tiles.")
 
         self.tiles = []
         if self.cache_dir:
@@ -153,7 +160,8 @@ class Generate_Tiled_Dataset(Dataset):
             self.cache_dir_lb.mkdir(parents=True, exist_ok=True)
 
         self.df = pd.DataFrame(
-            columns=["path_img", "path_lb", "is_empty", "bbx", "px_stats", "size"]
+            columns=["path_img", "path_lb",
+                     "is_empty", "bbx", "px_stats", "size"]
         )
         self.path_df = self.cache_dir / (self.name_data + ".csv")
 
@@ -204,10 +212,11 @@ class Generate_Tiled_Dataset(Dataset):
 
         if len(bbox) > 0:
             x, y, w, h = bbox
-            mask = mask[y : y + h, x : x + w]
+            mask = mask[y: y + h, x: x + w]
 
         cv2.imwrite(str(lb_path), mask)
-        self.df.loc[idx, :] = [img_path, lb_path, is_empty, bbox, px_stats, size]
+        self.df.loc[idx, :] = [img_path, lb_path,
+                               is_empty, bbox, px_stats, size]
         return (
             str(img_fpath),
             rle,
